pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "KohakuTerrarium"
version = "0.1.0"
description = "Terrarium: A universal agent framework for building any type of fully self-driven agent system"
authors = [
    { name = "Shih-Ying Yeh(KohakuBlueLeaf)", email = "apolloyeh0123@gmail.com" },
]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "openai>=2.0.0",
    "pydantic>=2.0.0",
    "rich>=14.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "aiofiles>=24.0.0",
    "python-dotenv>=1.0.0",
    "tomli>=2.0.0; python_version<'3.11'",
]
urls = { "Homepage" = "https://github.com/KohakuBlueLeaf/KohakuTerrarium" }

[project.scripts]
kt = "kohakuterrarium.__main__:main"

[project.optional-dependencies]
tui = [
    "textual>=3.0.0",
]
web = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "black>=25.12.0",
]
discord = [
    "discord.py>=2.0.0",
    "Pillow>=10.0.0",
]
asr = [
    "openai-whisper>=20240930",
    "sounddevice>=0.4.0",
    "numpy>=1.20.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"kohakuterrarium.builtin_skills" = ["**/*.md"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
